ULong better than intting

+++ Idle_project/Idle_project/Form1.vb
@@ -1,10 +1,10 @@
 ﻿Public Class Form1
     Dim EXP As Double = 0              'experience, used as currency for leveling up, modifiers for EXP gained include: perception, monsters and (special bonus for endgame)
-    Dim LVL As Integer = 1              'Levels, used to increase amount of experience gained per click/autoclick
-    Dim DarkVisionLVL As Integer = 0    'Levels in skill: Darkvision, Used to increase EXP gained and to unlock next part of the story.
-    Dim DarkVisionCost As Integer = 1   'Cost to level up the skill darkvision. (Might make this an overall skill EXP thing)
-    Dim SkillPoints As Integer = 0      'Skill points, used to level up all skills.
-    Dim LVLUpCost As Integer = 5        'Cost for leveling up, calculated based on this calculation: 1 + LVL^(0,25 * LVL)
+    Dim LVL As ULong = 1              'Levels, used to increase amount of experience gained per click/autoclick
+    Dim DarkVisionLVL As ULong = 0    'Levels in skill: Darkvision, Used to increase EXP gained and to unlock next part of the story.
+    Dim DarkVisionCost As ULong = 1   'Cost to level up the skill darkvision. (Might make this an overall skill EXP thing)
+    Dim SkillPoints As ULong = 0      'Skill points, used to level up all skills.
+    Dim LVLUpCost As ULong = 5        'Cost for leveling up, calculated based on this calculation: 1 + LVL^(0,25 * LVL)
     Dim AutoClick1_cost As Double = 50  'Cost of autoclicker, going to be based on some mathemagics
     Dim AutoclickExp As Double = 0      'Unknown, Kris can probably explain what it's going to be used for, "no i cant" -kris
